# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Student:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  services.CreateStudentRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - phone
    type: object
...
  services.UpdateStudentRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - phone
    type: object
  utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@englishcourse.com
    name: Novendra
  description: RESTful API backend untuk pendaftaran kursus bahasa Inggris dengan
    clean layered architecture.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: English Course Registration API
  version: "1.0"
paths:
...
  /students:
    get:
      description: Mengembalikan daftar semua peserta kursus yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Student'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil daftar seluruh student
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Membuat data peserta kursus baru dengan validasi email unik
      parameters:
      - description: Payload pendaftaran student
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Student'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mendaftarkan student baru
      tags:
      - Students
  /students/{id}:
    delete:
      description: Menghapus data student berdasarkan ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Menghapus data student
      tags:
      - Students
    get:
      description: Mengembalikan data detail student berdasarkan ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Student'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengambil detail profil student
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: Memperbarui nama, email, atau no telepon student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload update student
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Student'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mengubah data profil student
      tags:
      - Students
...
swagger: "2.0"